# Avoid O(n²) in removeDuplicateGradient