@nthykier 2018-03-10
This replaces #170