@nthykier 2018-03-10
Thanks :) 

Re: Performance on a large number of distinct gradients: No, I do not have a test case for that lying around.  Nonetheless, assuming I do not add any bugs in the changeset, the performance for those should improve similarly provided that the total number of gradients.  If you see a speed regression, it will be on SVGs with a low number of gradient - but then the total cost of removeDuplicateGradients will become negligible anyway (as there is simply not a lot of work to do).

The changeset is an optimal CPU time/memory trade-off from an algorithm complexity point of view.  If you double the number of gradients:

 * Original code: The runtime of removeDuplicateGradients basically quad-doubles while memory doubles (i.e. runtime is O(n²) and memory is O(n))

 * Patched code: The runtime and memory usage of removeDuplicateGradients basically doubles (i.e. both are O(n)).  The base amount of memory per gradient is slightly higher compared to the original code.

As I understand it, this memory trade-off is not a concern for scour at the moment, given we load the entire XML into memory.