@Ede123 2018-03-10
Just tested it on unique gradients and it's indeed ~5 times faster in my testcase.
I asked because the original code bailed out early in many cases, so I was not sure wether the reduced complexity of your code weighed out the additional computational effort, but that does not seem to be a relevant factor for usual files.